Image tag with some examples
Definition and Usage
- The <img> tag is used to embed an image in an HTML page.
- Images are not technically inserted into a web page; images are linked to web pages.
- The <img> tag creates a holding space for the referenced image.

The <img> tag has two required attributes:
------------------------------------------------
src - Specifies the path to the image
alt - Specifies an alternate text for the image, if the image for some reason cannot be displayed
Note: Also, always specify the width and height of an image. If width and height are not specified, the page might flicker while the image loads.